+++ tests/test_main_window_functions/test_no_file.py
@@ -1,11 +1,7 @@
-import PySide6.QtWidgets as QtW
 
 import ScenarioGUI.global_settings as global_vars
-from ScenarioGUI.gui_classes.gui_combine_window import MainWindow
 
-from ..gui_structure_for_tests import GUI
-from ..result_creating_class_for_tests import ResultsClass, data_2_results
-from ..test_translations.translation_class import Translations
+from ..starting_closing_tests import close_tests, start_tests
 
 
 def test_no_load_save_file(qtbot):
@@ -19,13 +15,11 @@ def test_no_load_save_file(qtbot):
     """
 
     # init gui window
-    main_window = MainWindow(QtW.QMainWindow(), qtbot, GUI, Translations, result_creating_class=ResultsClass, data_2_results_function=data_2_results)
-    main_window.delete_backup()
-    main_window = MainWindow(QtW.QMainWindow(), qtbot, GUI, Translations, result_creating_class=ResultsClass, data_2_results_function=data_2_results)
+    main_window = start_tests(qtbot)
     # check if an import error has been raises with a wrong load file
     main_window._load_from_data(f"not_there.{global_vars.FILE_EXTENSION}")
     assert main_window.status_bar.label.text() == main_window.translations.no_file_selected[0]
     # check if the current error message is shown with a wrong save file/folder
     main_window._save_to_data(f"hello/not_there.{global_vars.FILE_EXTENSION}")
     assert main_window.status_bar.label.text() == main_window.translations.no_file_selected[main_window.gui_structure.option_language.get_value()[0]]
-    main_window.delete_backup()
+    close_tests(main_window, qtbot)