+++ tests/test_main_window_functions/test_push_button_change.py
@@ -1,11 +1,6 @@
 import PySide6.QtCore as QtC
-import PySide6.QtWidgets as QtW
 
-from ScenarioGUI.gui_classes.gui_combine_window import MainWindow
-
-from ..gui_structure_for_tests import GUI
-from ..result_creating_class_for_tests import ResultsClass, data_2_results
-from ..test_translations.translation_class import Translations
+from ..starting_closing_tests import close_tests, start_tests
 
 
 def test_push_button_layout_change(qtbot):
@@ -18,9 +13,7 @@ def test_push_button_layout_change(qtbot):
         bot for the GUI
     """
     # init gui window
-    main_window = MainWindow(QtW.QMainWindow(), qtbot, GUI, Translations, result_creating_class=ResultsClass, data_2_results_function=data_2_results)
-    main_window.delete_backup()
-    main_window = MainWindow(QtW.QMainWindow(), qtbot, GUI, Translations, result_creating_class=ResultsClass, data_2_results_function=data_2_results)
+    main_window = start_tests(qtbot)
     main_window.add_scenario()
     # check if the layout is the small one at the beginning
     for page_i in main_window.gui_structure.list_of_pages:
@@ -38,4 +31,4 @@ def test_push_button_layout_change(qtbot):
             assert page_i.button.iconSize() == main_window.size_b
             assert page_i.button.size() == main_window.size_push_s
 
-    main_window.delete_backup()
+    close_tests(main_window, qtbot)